Added Java modules by separating into Gradle sub-projects.
