+++ modules/base/gradle.properties
@@ -1,0 +1,1 @@
+module.name=base
