+++ modules/base/src/main/java/com/harleyoconnor/javautilities/convention/NamingConvention.java
