+++ modules/base/src/main/java/module-info.java
@@ -1,0 +1,11 @@
+/**
+ * @author Harley O'Connor
+ */
+module javautilities.base {
+    requires org.jetbrains.annotations;
+
+    exports com.harleyoconnor.javautilities.convention;
+    exports com.harleyoconnor.javautilities.io;
+    exports com.harleyoconnor.javautilities.math;
+    exports com.harleyoconnor.javautilities;
+}
