+++ modules/collection/build.gradle.kts
@@ -1,0 +1,3 @@
+dependencies {
+    implementation(project(":reflect"))
+}
