+++ modules/collection/gradle.properties
@@ -1,0 +1,1 @@
+module.name=collection
