+++ modules/collection/src/main/java/module-info.java
@@ -1,0 +1,9 @@
+/**
+ * @author Harley O'Connor
+ */
+module javautilities.collection {
+    requires javautilities.reflect;
+    requires org.jetbrains.annotations;
+
+    exports com.harleyoconnor.javautilities.collection;
+}
