+++ modules/function/src/main/java/com/harleyoconnor/javautilities/function/ThrowableTriFunction.java
