+++ modules/function/src/main/java/com/harleyoconnor/javautilities/function/TriFunction.java
