+++ modules/function/src/main/java/module-info.java
@@ -1,0 +1,8 @@
+/**
+ * @author Harley O'Connor
+ */
+module javautilities.function {
+    requires org.jetbrains.annotations;
+
+    exports com.harleyoconnor.javautilities.function;
+}
