+++ modules/reflect/build.gradle.kts
@@ -1,0 +1,4 @@
+dependencies {
+    implementation(project(":base"))
+    implementation(project(":function"))
+}
