+++ modules/reflect/src/main/java/module-info.java
@@ -1,0 +1,10 @@
+/**
+ * @author Harley O'Connor
+ */
+module javautilities.reflect {
+    requires javautilities.base;
+    requires javautilities.function;
+    requires org.jetbrains.annotations;
+
+    exports com.harleyoconnor.javautilities.reflect;
+}
