--- src/main/java/module-info.java
@@ -1,13 +1,0 @@
-/**
- * @author Harley O'Connor
- */
-module JavaUtilities.main {
-    requires org.jetbrains.annotations;
-
-    exports com.harleyoconnor.javautilities.collection;
-    exports com.harleyoconnor.javautilities.convention;
-    exports com.harleyoconnor.javautilities.function;
-    exports com.harleyoconnor.javautilities.pair;
-    exports com.harleyoconnor.javautilities.reflect;
-    exports com.harleyoconnor.javautilities.util;
-}
